Replace sigpause() with pause()
Since xgks was written, signals have been standardized by POSIX
with a different API but similar functionality.  The intent of
sigpause() in this code was to wait for some signal to fire
and then continue the main loop after the signal has been
processed.  Replace calls to sigpause() with a call to pause()
from <unistd.h> that has the same effect.